New version of Zenzero - 1.6.3

+++ footer.php
@@ -9,22 +9,23 @@
 ?>
 
 	</div><!-- #content -->
-
-	<footer id="colophon" class="site-footer">
-		<div class="site-info smallPart">
-			<?php $zenzero_copyrightText = get_theme_mod('zenzero_theme_options_copyright', '&copy; '.date('Y').' '. get_bloginfo('name')); ?>
-			<?php echo wp_kses($zenzero_copyrightText, zenzero_allowed_html()); ?>
-			<span class="sep"> | </span>
-			<?php
-			/* translators: 1: theme name, 2: theme developer */
-			printf( esc_html__( 'WordPress Theme: %1$s by %2$s.', 'zenzero' ), '<a target="_blank" href="https://crestaproject.com/downloads/zenzero/" rel="nofollow" title="Zenzero Theme">Zenzero</a>', 'CrestaProject' );
+	<?php if ( ! function_exists( 'elementor_theme_do_location' ) || ! elementor_theme_do_location( 'footer' ) ) : ?>
+		<footer id="colophon" class="site-footer">
+			<div class="site-info smallPart">
+				<?php $zenzero_copyrightText = get_theme_mod('zenzero_theme_options_copyright', '&copy; '.date('Y').' '. get_bloginfo('name')); ?>
+				<?php echo wp_kses($zenzero_copyrightText, zenzero_allowed_html()); ?>
+				<span class="sep"> | </span>
+				<?php
+				/* translators: 1: theme name, 2: theme developer */
+				printf( esc_html__( 'WordPress Theme: %1$s by %2$s.', 'zenzero' ), '<a target="_blank" href="https://crestaproject.com/downloads/zenzero/" rel="nofollow" title="Zenzero Theme">Zenzero</a>', 'CrestaProject' );
+				?>
+			</div><!-- .site-info -->
+			<?php 
+			$zenzero_hideSearch = get_theme_mod('zenzero_theme_options_hidesearch', '1');
+			zenzero_social_button();
 			?>
-		</div><!-- .site-info -->
-		<?php 
-		$zenzero_hideSearch = get_theme_mod('zenzero_theme_options_hidesearch', '1');
-		zenzero_social_button();
-		?>
-	</footer><!-- #colophon -->
+		</footer><!-- #colophon -->
+	<?php endif; ?>
 </div><!-- #page -->
 <?php get_sidebar(); ?>
 <a href="#top" id="toTop" class="showTop"><i class="fa fa-angle-up"></i></a>
